Web date range: use short month names so it stays on one line
renderMetaRange's date formatter used { month: 'long' } ("September 14"),
which was wide enough to wrap the range onto two lines. Switch to
{ month: 'short' } ("Sep 14 – Sep 30, 2026") — verified single-line at
the dock panel's width. Rebuilt standalone.html.

+++ planner/app.js
@@ -2812,7 +2812,7 @@
     renderMetaRange(trip) {
       const parse = s => { const dt = new Date((s || '') + 'T00:00:00'); return isNaN(dt.getTime()) ? null : dt; };
       const dep = parse(trip.depart), ret = parse(trip.returnDate);
-      const md = dt => dt ? dt.toLocaleDateString('en-US', { month: 'long', day: 'numeric' }) : 'Add date';
+      const md = dt => dt ? dt.toLocaleDateString('en-US', { month: 'short', day: 'numeric' }) : 'Add date';
       const sameYear = dep && ret && dep.getFullYear() === ret.getFullYear();
       const withYear = (dt, base) => (dep && ret && !sameYear && dt) ? `${base}, ${dt.getFullYear()}` : base;
       const depTxt = withYear(dep, md(dep)), retTxt = withYear(ret, md(ret));
